Update BriemHand.glyphs
Fixed one malformed join in black. Switched one join to look better, tweaked a join. Removed some bad kerning. adjusted a diacric anchor location and shape. Added a default VF origin and axis location values to exports. Note: The VF is menuing in the TN tool not as Breim hand but Breim Hand light. Not sure if that is an issue.