Update BriemHand.glyphs
Corrected some diacritic placement errors. Added code to handle the zigzag made by typing " /|/|/ " etc including forcing an outstroke at the end. Corrected a join that Briem wanted switched from a.a to a.b. Turned p joins back on again. The iacute+z problem doesn't look like it was in this file and the iacute+i isn't previewing in glyphs so I'm going to see if a build made here can be tested. Briem also says "Some sidebearings are inconsistent as well." This is vague. Similarly "Some lettershapes have mutated. But that is an undertaking for another day." Utterly vague. The issue listed with oslash is a false report from what I can see. Acute positions fixed.